# Conflicts:
#	src/main/java/de/dafuqs/spectrum/recipe/pedestal/BuiltinGemstoneColor.java

+++ src/main/java/de/dafuqs/spectrum/items/tooltip/BottomlessBundleTooltipData.java
@@ -3,6 +3,6 @@ package de.dafuqs.spectrum.items.tooltip;
 import net.minecraft.item.*;
 import net.minecraft.item.tooltip.TooltipData;
 
-public record BottomlessBundleTooltipData(ItemStack itemStack, int amount) implements TooltipData {
+public record BottomlessBundleTooltipData(ItemStack itemStack, long amount) implements TooltipData {
 
 }
